Extracting outputs from bayes_met() objects

This function extracts outputs of the Bayesian model fitted using bayes_met(), and provides some diagnostics about the model

extr_outs(
  data,
  trait,
  model,
  probs = c(0.025, 0.975),
  check.stan.diag = TRUE,
  verbose = FALSE,
  ...
)

Arguments

data

A data frame containing the observations

trait

A character representing the name of the column that corresponds to the analysed trait

model

An object containing the Bayesian model fitted using rstan

probs

A vector with two elements representing the probabilities (in decimal scale) that will be considered for computing the quantiles.

check.stan.diag

A logical value indicating whether the function should extract some diagnostic using native rstan functions.

verbose

A logical value. If TRUE, the function will indicate the completed steps. Defaults to FALSE

...

Passed to rstan::stan_diag()

Value

The function returns a list with:

post : a list with the posterior of the effects, and the data generated by the model

map : a list with the maximum posterior values of each effect

ppcheck : a matrix containing the p-values of maximum, minimum, median, mean and standard deviation; effective number of parameters, WAIC2 value, Rhat and effective sample size.

plots : a list with three types of ggplots: histograms, trace plots and density plots. These will be available for all effects declared at the effects argument.

stan_plots: If check.stan.diag = TRUE, a list with plots generated by rstan::stan_diag()

Details

More details about the usage of extr_outs, as well as the other function of the ProbBreed package can be found at https://saulo-chaves.github.io/ProbBreed_site/.

See also

Examples

# \donttest{
mod = bayes_met(data = maize,
                gen = "Hybrid",
                loc = "Location",
                repl = c("Rep", "Block"),
                year = NULL,
                reg = 'Region',
                res.het = FALSE,
                trait = 'GY',
                iter = 6000, cores = 4, chains = 4)
#> Error in rstan::stan_model(model_code = stan_df, model_name = "BayesMET"): Boost not found; call install.packages('BH')

outs = extr_outs(data = maize, trait = "GY", model = mod,
                 probs = c(0.05, 0.95),
                 check.stan.diag = TRUE,
                 verbose = TRUE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'extract': object 'mod' not found
                 # }