Maize real data set

This dataset belongs to value of cultivation and use maize trials of Embrapa Maize and Sorghum, and was used by Dias et al. (2022). It contains the grain yield of 32 single-cross hybrids and four commercial checks (36 genotypes in total) evaluated in 16 locations across five regions or mega-environments. These trials were laid out in incomplete blocks design, using a block size of 6 and two replications per trial.

Usage

maize

Format

maize

A data frame with 823 rows and 6 columns:

Location

16 locations

Region

5 regions

Rep

2 replicates

Block

6 blocks

Hybrid

36 genotypes

GY

Grain yield (phenotypes)

Source

Dias, K. O. G, Santos J. P. R., Krause, M. D., Piepho H. -P., Guimarães, L. J. M., Pastina, M. M., and Garcia, A. A. F. (2022). Leveraging probability concepts for cultivar recommendation in multi-environment trials. Theoretical and Applied Genetics, 133(2):443-455. \Sexpr[results=rd]{tools:::Rd_expr_doi("#1")}10.1007/s00122-022-04041-ytools:::Rd_expr_doi("10.1007/s00122-022-04041-y")