Plots for the extr object

Build plots using the outputs stored in the extr object.

Usage

# S3 method for class 'extr'
plot(x, ..., category = "ppdensity")

Arguments

x

An object of class extr.

...

Passed to ggplot2::geom_histogram, when category = histogram. Useful to change the number of bins.

category

A string indicating which plot to build. See options in the Details section.

Details

The available options are:

ppdensity : Density plots of the empirical and sampled data, useful to assess the model's convergence.

density : Density plots of the model's effects.

histogram : Histograms of the model's effects.

traceplot: Trace plot showing the changes in the effects' values across iterations and chains.

See also

Examples

# \donttest{
mod = bayes_met(data = soy,
                gen = "Gen",
                loc = "Loc",
                repl = NULL,
                year = NULL,
                reg = NULL,
                res.het = FALSE,
                trait = 'Y',
                iter = 6000, cores = 1, chains = 4)
#> Error in bayes_met(data = soy, gen = "Gen", loc = "Loc", repl = NULL,     year = NULL, reg = NULL, res.het = FALSE, trait = "Y", iter = 6000,     cores = 1, chains = 4): By default, a model with no replicates can only be fitted with heterogeneous residual variances (res.het = TRUE)

outs = extr_outs(model = mod,
                 probs = c(0.05, 0.95),
                 verbose = TRUE)
#> Error in eval(expr, envir, enclos): object 'mod' not found
plot(results, category = "ppdensity")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "density")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "histogram")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "traceplot")
#> Error in eval(expr, envir, enclos): object 'results' not found
# }