Plots for the probsup object

Build plots using the outputs stored in the probsup object.

# S3 method for class 'probsup'
plot(obj, category = "perfo", level = "across", ...)

Arguments

obj

An object of class probsup.

category

A string indicating which plot to build. See options in the Details section.

level

A string indicating the information level to be used for building the plots. Options are "across" for focusing on the probabilities across environments, or "within" to focus on the within-environment effects. Defaults to "across".

...

currently not used

Details

The available options are:

hpd : a caterpillar plot representing the marginal genotypic value of each genotype, and their respective highest posterior density interval (95% represented by the thick line, and 97.5% represented by the thin line). Available only if level = "across".

perfo : if level = "across", a lollipop plot illustrating the probabilities of superior performance. If level = "within", a heatmap with the probabilities of superior performance within environments. If a model with reg and/or year is fitted, multiple plots are produced.

stabi: a lollipop plot with the probabilities of superior stability. If a model with reg and/or year is fitted, multiple plots are produced. Available only if level = "across".

pair_perfo : if level = "across", a heatmap representing the pairwise probability of superior performance (the probability of genotypes at the x-axis being superior to those on the y-axis). If level = "within", a list of heatmaps representing the pairwise probability of superior performance within environments. If a model with reg and/or year is fitted, multiple plots (and multiple lists) are produced. Should this option is set, it is mandatory to store the outputs in an object (e.g., pl <- plot(obj, category = "pair_perfo", level = "within")) so they can be visualized one at a time.

pair_stabi: a heatmap with the pairwise probabilities of superior stability (the probability of genotypes at the x-axis being more stable than those on the y-axis). If a model with reg and/or year is fitted, multiple plots are produced. Available only if level = "across".

joint: a lollipop plot with the joint probabilities of superior performance and stability.

See also

Examples

mod = bayes_met(data = soy,
                gen = "Gen",
                loc = "Loc",
                repl = NULL,
                year = NULL,
                reg = NULL,
                res.het = FALSE,
                trait = 'Y',
                iter = 6000, cores = 4, chains = 4)
#> Error in bayes_met(data = soy, gen = "Gen", loc = "Loc", repl = NULL,     year = NULL, reg = NULL, res.het = FALSE, trait = "Y", iter = 6000,     cores = 4, chains = 4): By default, a model with no replicates can only be fitted with heterogeneous residual variances (res.het = TRUE)

outs = extr_outs(model = mod,
                 probs = c(0.05, 0.95),
                 verbose = TRUE)
#> Error in eval(expr, envir, enclos): object 'mod' not found

results = prob_sup(extr = outs,
                   int = .2,
                   increase = TRUE,
                   save.df = FALSE,
                   verbose = FALSE)
#> Error in eval(expr, envir, enclos): object 'outs' not found

plot(results, category = "hpd")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "perfo", level = "across")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "perfo", level = "within")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "stabi")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "pair_perfo", level = "across")
#> Error in eval(expr, envir, enclos): object 'results' not found
plwithin = plot(results, category = "pair_perfo", level = "within")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "pair_stabi")
#> Error in eval(expr, envir, enclos): object 'results' not found
plot(results, category = "joint")
#> Error in eval(expr, envir, enclos): object 'results' not found