Probabilities of superior performance and stability

This function estimates the probabilities of superior performance and stability across environments (marginal output). It also computes the probabilities of superior performance within environments (conditional output).

prob_sup(
  data,
  trait,
  gen,
  env,
  reg = NULL,
  extr_outs,
  int,
  increase = TRUE,
  save.df = FALSE,
  interactive = FALSE
)

Arguments

data

A data frame containing the phenotypic data

trait, gen, env

A string. The name of the columns that corresponds to the variable, genotype and environment information, respectively

reg

A string or NULL. If the data set has information about regions, reg will be a string with the name of the column that corresponds to the region information. Otherwise, reg = NULL (default).

extr_outs

An object from the extr_outs() function

int

A number representing the selection intensity (between 0 and 1)

increase

Logical. Indicates the direction of the selection. TRUE (default) for increasing the trait value, FALSE otherwise.

save.df

Logical. Should the data frames be saved in the work directory? TRUE for saving, FALSE (default) otherwise.

interactive

Logical. Should ggplots be converted into interactive plots? If TRUE, the function loads the plotly package and uses the plotly::ggplotly() command.

Value

The function returns two lists, one with the marginal probabilities, and another with the conditional probabilities.

The marginal list has:

df : A list of data frames containing the calculated probabilities:

perfo: the probabilities of superior performance

pair_perfo: the pairwise probabilities of superior performance

stabi: the probabilities of superior stability. When reg is not NULL, stabi is divided into stabi_gl for the stability across environments, and stabi_gm for the stability across regions

pair_stabi: the pairwise probabilities of superior stability, which is also divided into pair_stabi_gl and pair_stabi_gm when reg is not NULL

joint_prob: the joint probabilities of superior performance and stability

plot : A list of ggplots illustrating the outputs:

g_hpd: a caterpillar plot representing the marginal genetic value of each genotype, and their respective high posterior density (95% represented by the thick line, and 97.5% represented by the thin line)

perfo: a bar plot illustrating the probabilities of superior performance

pair_perfo: a heatmap representing the pairwise probability of superior performance (the probability of genotypes at the x-axis being superior to those on y-axis)

stabi: a bar plot with the probabilities of superior stability. Like the data frames, when reg is not NULL, two different plots are generated, one for the stability across environments (stabi_gl), and another for the stability across regions (stabi_gm)

pair_stabi: a heatmap with the pairwise probabilities of superior stability (also divided into two different plots when reg is not NULL). This plot represents the probability of genotypes at the x-axis being superior to those on y-axis

joint_prob: a plot with the probabilities of superior performance, probabilities of superior stability and the joint probabilities of superior performance and stability.

The conditional list has:

df : A list with:

perfo: a data frame containing the probabilities of superior performance within environments. It also has the probabilities of superior performance with regions if reg is not NULL.

pair_perfo: a list with the pairwise probabilities of superior performance within environments. If reg is not NULL, two lists are generated.

plot : A list with:

perfo: a heatmap with the probabilities of superior performance within environments. If reg is not NULL, there will be two heatmaps: perfo_env for the probabilities of superior performance within environments, and perfo_reg for the same probabilities within regions.

pair_perfo: a list of heatmaps representing the pairwise probability of superior performance within environments. If reg is not NULL, there will be two lists: pair_perfo_env for the probabilities of superior performance within environments, and pair_perfo_reg for the same probabilities within regions. The interpretation is the same as in the pair_perfo in the marginal list: the probability of genotypes at the x-axis being superior to those on y-axis.

Details

Probabilities provide the risk of recommending a candidate for a target population of environments or for a specific environment. The function prob_sup() computes the probabilities of superior performance (agronomic stability, or previsibility), and the probabilities of superior stability (ecological stability, or invariance):

Probability of superior performance

Let \(\Omega\) represent the subset of selected genotypes based on their performance across environments. A given genotype \(j\) will belong to \(\Omega\) if its genetic marginal value (\(g_j\)) is among the top \(V\), with \(V\) being the selection intensity. prob_sup() leverages the Monte Carlo discretized sampling from the posterior distribution to emulate the occurrence of \(S\) trials. Thence, the probability of the \(j^{th}\) genotype belonging to \(\Omega\) is the ratio of success (\(g_j \in \Omega\)) and the total number of sampled events (the emulated \(S\) trials):

$$Pr(g_j \in \Omega \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_j^{(s)} \in \Omega \vert y)}$$

where \(S\) is the total number of samples (\(s = 1, 2, ..., S\)), and \(I(g_j^{(s)} \in \Omega \vert y)\) is an indicator variable mapping success (1) if \(g_j^{(s)}\) exists in \(\Omega\) in the \(s^{th}\) sample, and failure (0) otherwise. \(S\) is conditioned to the number of iterations and chains previously set at bayes_met().

The same idea can be applied for each environment: let \(\Omega_k\) represent the subset of superior genotypes in the \(k^{th}\) environment, then the probability of the \(j^{th}\) genotype belonging to \(\Omega_k\) is

$$Pr(g_{jk} \in \Omega_k \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_{jk}^{(s)} \in \Omega_k \vert y)}$$

where \(I(g_{jk}^{(s)} \in \Omega_k \vert y)\) is an indicator variable mapping success (1) if \(g_{jk}^{(s)}\) exists in \(\Omega_k\), and failure (0) otherwise. Note that \(g_{jk} = g_j + ge_{jk}\), i.e., the genetic marginal value is summed to the genetic value of \(j\) when planted specifically at environment \(k\) (genotype-by-environment interaction effect).

Pairwise comparisons are also available in prob_sup(), both across and within environments. Leveraging the same samples, prob_sup() estimates the probability of a genotype \(j\) performing better than a genotype \(i\). Across environments, the equation is the following:

$$Pr(g_j > g_i \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_j^{(s)} > g_i^{(s)} \vert y)}$$

or

$$Pr(g_j < g_i \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_j^{(s)} < g_i^{(s)} \vert y)}$$

The first equation is applied when the aim is to increase the trait value (increase = T). Conversely, the second equation is used when increase = F. Within environments, \(g_j\) is switched by \(g_{jk}\).

Probability of superior stability

The equations described above are related to the genotypes' previsibility, an agronomic way of looking at stability. Another point of view is to consider the invariance as a measure of stability. Making a direct analogy with Shukla's (1972) theory, a stable genotype would be the one that had the lower variance of the genotype-by-environment interaction effect, \(var(ge_{jk})\). Using the same ideas as above, the probability of superior stability is given by:

$$Pr [var(ge_{jk})\in \Omega \vert y] = \frac{1}{S} \sum_{s=1}^S{I [var(ge_{jk}^{(s)})\in \Omega \vert y]} $$

where \(I[var(ge_{jk}^{(s)})\in \Omega \vert y]\) is an indicator variable mappings success (1) or failure (0).

prob_sup() also provides means to directly compare the stability of two genotypes. The pairwise probability of superior stability is given by:

$$Pr[var(ge_{jk}) < var(ge_{ik}) \vert y] = \frac{1}{S} \sum_{s=1}^S{I [var(ge_{jk}^{(s)}) < var(ge_{ik}^{(s)}) \vert y]} $$

Note that \(j\) will be superior to \(i\) if it has a lower variance of the genotype-by-environment interaction effect. This is true regardless if increase is set to TRUE or FALSE.

Finally, by assuming independence between the genetic main effect and the genotype-by-environment effect variance, prob_sup computes the joint probability of superior performance and stability:

$$Pr[g_j \in \Omega, var(ge_{jk} \in \Omega)] = Pr(g_j \in \Omega \vert y) \times Pr[var(ge_{jk}) \in \Omega \vert y]$$

The estimation of these probabilities are strictly related to some key question that constantly arises in plant breeding:

What is the risk of recommending a selection candidate for a target population of environments?

What is the probability of a given selection candidate having good performance if recommended to a target population of environments? And for a specific environment?

What is the probability of a given selection candidate having better performance than a cultivar check in the target population of environments? And in specific environments?

How probable is it that a given selection candidate performs similarly across environments?

What are the chances that a given selection candidate is more stable than a cultivar check in the target population of environments?

What is the probability that a given selection candidate having a superior and invariable performance across environments?

References

Dias, K. O. G, Santos J. P. R., Krause, M. D., Piepho H. -P., Guimarães, L. J. M., Pastina, M. M., and Garcia, A. A. F. (2022). Leveraging probability concepts for cultivar recommendation in multi-environment trials. Theoretical and Applied Genetics, 133(2):443-455. https://doi.org/10.1007/s00122-022-04041-y

Shukla, G. K. (1972) Some statistical aspects of partioning genotype environmental componentes of variability. Heredity, 29:237-245. https://doi.org/10.1038/hdy.1972.87

Examples

if (FALSE) {
mod = bayes_met(data = soy,
                gen = "Gen",
                env = "Env",
                repl = NULL,
                reg = "Reg",
                res.het = F,
                trait = "Y",
                iter = 2000, cores = 1, chains = 4)

outs = extr_outs(data = soy, trait = "Y", gen = "Gen", model = mod,
                 effects = c('l','g','gl','m','gm'),
                 nenv = length(unique(soy$Env)), res.het = FALSE,
                 probs = c(0.05, 0.95), check.stan.diag = TRUE)

results = prob_sup(data = soy, trait = "Y", gen = "Gen", env = "Env",
                   extr_outs = outs, reg = 'Reg', int = .2,
                   increase = T, save.df = T, interactive = T)
}