Factor recoding

The functions of ProbBreed works with alphanumeric factors (for e.g., "G01"). recod was built to perform the recoding of a factor's levels in the dataset. It adds a new column with the newly coded levels, keeping the original names for reference.

recod(data, name.fact, cod)

Arguments

data

A data frame containing the observations

name.fact

A string. The name of the column containing the factor to be recoded

cod

A string. The letter that will appear before the numbers in the new code. For e.g.: "G01"

Value

The function returns the same data frame, but with a new column with the prefix "cod." containing the recoded levels

Examples

toy = data.frame(
            gen = rep(1:5, each =4),
            env = rep(1:4, times = 5),
            y = rnorm(20, mean = 20, sd = 5)
        )
new.toy = recod(toy, name.fact = 'gen', cod = 'G')
toy
#>    gen env        y
#> 1    1   1 20.33643
#> 2    1   2 17.41481
#> 3    1   3 20.10161
#> 4    1   4 17.69680
#> 5    2   1 24.36701
#> 6    2   2 22.81026
#> 7    2   3 28.78699
#> 8    2   4 26.86547
#> 9    3   1 19.75205
#> 10   3   2 27.79512
#> 11   3   3 16.45833
#> 12   3   4 20.33524
#> 13   4   1 19.92416
#> 14   4   2 18.44004
#> 15   4   3 22.88189
#> 16   4   4 23.80263
#> 17   5   1 15.47812
#> 18   5   2 26.51771
#> 19   5   3 16.29485
#> 20   5   4 19.63907