Bayesian model for multi-environment trials

This function runs a Bayesian model for analysing data from Multi-environment trials using rstan, the R interface to Stan.

bayes_met(
  data,
  gen,
  env,
  repl,
  trait,
  reg = NULL,
  res.het = F,
  iter = 2000,
  cores = 1,
  chains = 4,
  ...
)

Arguments

data

A data frame containing the observations

gen, env

A string. The name of the column that corresponds to the evaluated genotype or environment.

repl

A string, a vector, or NULL. If the trial is randomized in complete blocks, repl will be a string representing the name of the column that corresponds to the blocks. If the trial is randomized in incomplete blocks design, repl will be a string vector containing the name of the column that corresponds to the replicate and block effects on the first and second positions, respectively. If the data do not have replicates, repl will be NULL.

trait

A string. The name of the column that corresponds to the analysed variable.

reg

A string or NULL. If the data set has information about regions, reg will be a string with the name of the column that corresponds to the region information. Otherwise, reg = NULL (default).

res.het

Logical, indicating if the model should consider heterogeneous residual variances. Default is FALSE.

iter

Inherited from rstan::sampling(). A positive integer specifying the number of iterations for each chains (including warmup). The default is 2000.

cores

Inherited from rstan::sampling(). A positive integer specifying the number of cores to use when executing the chains in parallel (defaults to 1).

chains

Inherited from rstan::sampling(). A positive integer specifying the number of Markov chains. The default is 4.

...

Additional arguments passed to the rstan::sampling() function (for instance, to change the thin number, or to set a specific seed). For more information, see rstan::sampling() manual

Value

The function returns an object of S4 class stanfit containing the fitted results

See also

Examples

if (FALSE) {
# Model
mod = bayes_met(data = soy,
                gen = "Gen",
                env = "Env",
                repl = NULL,
                reg = "Reg",
                res.het = F,
                trait = "Y",
                iter = 2000, cores = 1, chains = 4)
                }