Extracting outputs from bayes_met() objects

This function extracts outputs of the Bayesian model fitted using bayes_met(), and provides some diagnostics about the model

extr_outs(
  data,
  trait,
  gen,
  model,
  effects,
  nenv,
  res.het = FALSE,
  probs = c(0.025, 0.975),
  check.stan.diag = FALSE,
  ...
)

Arguments

data

A data frame containing the observations

trait

A character representing the name of the column that corresponds to the analysed trait

gen

A character representing the name of the column that corresponds to the evaluated genotypes

model

An object containing the Bayesian model fitted using rstan

effects

A string vector containing the codes of the effects included in the model. The codes are:

r : replicate effect

b : block inside replicate effect (if the trials are in incomplete blocks design)

l : environment effect

m : region effect (if it is available)

g : genetic effect

gl : genotype-by-environment effect

gm : genotype-by-region effect

nenv

The number of environments in the analysis

res.het

Are the residual heterogeneous? Default is FALSE

probs

A vector with two elements representing the probabilities that will be considered for computing the quantiles. Provide values between 0 and 1.

check.stan.diag

A logical value indicating whether the function should extract some diagnostic using native rstan functions.

...

Passed to rstan::stan_diag()

Value

The function returns a list with:

post : a list with the posterior of the effects, and the data generated by the model

variances : a data frame with the variances of each effect

map : a list with the maximum posterior values of each effect

ppcheck : a matrix containing the p-values of maximum, minimum, median, mean and standard deviation; effective number of parameters, WAIC2 value and Rhat

fun.plots : a list with three types of ggplots: histograms, trace plots and density plots. These will be available for all effects declared at the effects argument.

See also

Examples

if (FALSE) {
mod = bayes_met(data = soy,
                gen = "Gen",
                env = "Env",
                repl = NULL,
                reg = "Reg",
                res.het = F,
                trait = "Y",
                iter = 100, cores = 4, chain = 4)
                #Remember, increase the number of iterations, cores and chains

outs = extr_outs(data = soy, trait = "Y", gen = "Gen", model = mod,
                 effects = c('l','g','gl','m','gm'),
                 nenv = length(unique(soy$Env)), res.het = FALSE,
                 probs = c(0.05, 0.95), check.stan.diag = TRUE)
                 }