Extract outputs from stanfit objects obtained from bayes_met

Extracts outputs of the Bayesian model fitted using bayes_met(), and provides some diagnostics.

extr_outs(model, probs = c(0.025, 0.975), verbose = FALSE)

Arguments

model

An object of class stanfit, obtained using bayes_met()

probs

A vector with two elements representing the probabilities (in decimal scale) that will be considered for computing the quantiles.

verbose

A logical value. If TRUE, the function will indicate the completed steps. Defaults to FALSE

Value

The function returns an object of class extr, which is a list with:

variances : a data frame containing the variance components of the model effects, their standard deviation, naive standard error and highest posterior density interval.

post : a list with the posterior of the effects, and the data generated by the model.

map : a list with the maximum posterior values of each effect

ppcheck : a matrix containing the p-values of maximum, minimum, median, mean and standard deviation; effective number of parameters, WAIC2 value, Rhat and effective sample size.

Details

More details about the usage of extr_outs and other functions of the ProbBreed package can be found at https://saulo-chaves.github.io/ProbBreed_site/.

See also

Examples

mod = bayes_met(data = soy,
                gen = "Gen",
                loc = "Loc",
                repl = NULL,
                year = NULL,
                reg = NULL,
                res.het = TRUE,
                trait = 'Y',
                iter = 6000, cores = 4, chains = 4)
#> Error in rstan::stan_model(model_code = stan_df, model_name = "BayesMET"): Boost not found; call install.packages('BH')

outs = extr_outs(model = mod,
                 probs = c(0.05, 0.95),
                 verbose = TRUE)
#> Error in eval(expr, envir, enclos): object 'mod' not found