Extract outputs from stanfit objects obtained from bayes_met

Extracts outputs of the Bayesian model fitted using bayes_met(), and provides some diagnostics.

Usage

extr_outs(model, probs = c(0.025, 0.975), verbose = FALSE)

Arguments

model

An object of class stanfit, obtained using bayes_met()

probs

A vector with two elements representing the probabilities (in decimal scale) that will be considered for computing the quantiles.

verbose

A logical value. If TRUE, the function will indicate the completed steps. Defaults to FALSE

Value

The function returns an object of class extr, which is a list with:

variances : a data frame containing the variance components of the model effects, their standard deviation, naive standard error and highest posterior density interval.

post : a list with the posterior of the effects, and the data generated by the model.

map : a list with the maximum posterior values of each effect

ppcheck : a matrix containing the p-values of maximum, minimum, median, mean and standard deviation; effective number of parameters, WAIC2 value, Rhat and effective sample size.

Details

More details about the usage of extr_outs and other functions of the ProbBreed package can be found at https://saulo-chaves.github.io/ProbBreed_site/.

See also

Examples

# \donttest{
mod = bayes_met(data = soy,
                gen = "Gen",
                loc = "Loc",
                repl = NULL,
                year = NULL,
                reg = NULL,
                res.het = TRUE,
                trait = 'Y',
                iter = 6000, cores = 1, chains = 4)
#> 
#> SAMPLING FOR MODEL 'BayesMET' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000118 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.18 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 6000 [  0%]  (Warmup)
#> Chain 1: Iteration:  600 / 6000 [ 10%]  (Warmup)
#> Chain 1: Iteration: 1200 / 6000 [ 20%]  (Warmup)
#> Chain 1: Iteration: 1800 / 6000 [ 30%]  (Warmup)
#> Chain 1: Iteration: 2400 / 6000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 3000 / 6000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 3001 / 6000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 3600 / 6000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 4200 / 6000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 4800 / 6000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 5400 / 6000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 6000 / 6000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 29.648 seconds (Warm-up)
#> Chain 1:                26.769 seconds (Sampling)
#> Chain 1:                56.417 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'BayesMET' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 8.7e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.87 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 6000 [  0%]  (Warmup)
#> Chain 2: Iteration:  600 / 6000 [ 10%]  (Warmup)
#> Chain 2: Iteration: 1200 / 6000 [ 20%]  (Warmup)
#> Chain 2: Iteration: 1800 / 6000 [ 30%]  (Warmup)
#> Chain 2: Iteration: 2400 / 6000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 3000 / 6000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 3001 / 6000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 3600 / 6000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 4200 / 6000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 4800 / 6000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 5400 / 6000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 6000 / 6000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 25.109 seconds (Warm-up)
#> Chain 2:                26.636 seconds (Sampling)
#> Chain 2:                51.745 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'BayesMET' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 8.6e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.86 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 6000 [  0%]  (Warmup)
#> Chain 3: Iteration:  600 / 6000 [ 10%]  (Warmup)
#> Chain 3: Iteration: 1200 / 6000 [ 20%]  (Warmup)
#> Chain 3: Iteration: 1800 / 6000 [ 30%]  (Warmup)
#> Chain 3: Iteration: 2400 / 6000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 3000 / 6000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 3001 / 6000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 3600 / 6000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 4200 / 6000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 4800 / 6000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 5400 / 6000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 6000 / 6000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 27.487 seconds (Warm-up)
#> Chain 3:                26.556 seconds (Sampling)
#> Chain 3:                54.043 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'BayesMET' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000103 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.03 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 6000 [  0%]  (Warmup)
#> Chain 4: Iteration:  600 / 6000 [ 10%]  (Warmup)
#> Chain 4: Iteration: 1200 / 6000 [ 20%]  (Warmup)
#> Chain 4: Iteration: 1800 / 6000 [ 30%]  (Warmup)
#> Chain 4: Iteration: 2400 / 6000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 3000 / 6000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 3001 / 6000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 3600 / 6000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 4200 / 6000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 4800 / 6000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 5400 / 6000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 6000 / 6000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 27.458 seconds (Warm-up)
#> Chain 4:                26.526 seconds (Sampling)
#> Chain 4:                53.984 seconds (Total)
#> Chain 4: 

outs = extr_outs(model = mod,
                 probs = c(0.05, 0.95),
                 verbose = TRUE)
#> -> Posterior effects extracted
#> -> Variances extracted
#> -> Maximum posterior values extracted
#> -> Posterior predictive checks computed
#> 
#> Divergences:
#> 0 of 12000 iterations ended with a divergence.
#> 
#> Tree depth:
#> 0 of 12000 iterations saturated the maximum tree depth of 10.
#> 
#> Energy:
#> E-BFMI indicated no pathological behavior.
# }